Divided graphs over multiple pages. Each page has its own graph
