+++ dashboard/routings/__init__.py
@@ -11,6 +11,7 @@ import dashboard.routings.login
 import dashboard.routings.setup
 import dashboard.routings.result
 import dashboard.routings.export_data
+import dashboard.routings.measurements
 
 # Provide a secret-key for using WTF-forms
 if user_app.secret_key is None:
