--- dashboard/templates/endpoint.html
@@ -1,105 +1,0 @@
-{% extends "main.html" %}
-{% block content %}
-
-    <div class="row">
-        <!-- Advanced Tables -->
-        <div class="panel panel-default">
-            <div class="panel-heading">
-                Summary of endpoint
-            </div>
-            <br/>
-            <div class="panel-body">
-                <div class="table-responsive">
-                    <table class="table table-striped table-bordered table-hover">
-                        <tr>
-                            <td><b>Endpoint</b></td>
-                            <td>{{ rule.endpoint }}</td>
-                        </tr>
-                        <tr>
-                            <td><b>Added since app version</b></td>
-                            <td>{{ rule.version_added }}</td>
-                        </tr>
-                        <tr>
-                            <td><b>Date added to app</b></td>
-                            <td>{{ "{:%Y-%m-%d %H:%M:%S }".format(rule.time_added) }}</td>
-                        </tr>
-                        {% if url %}
-                            <tr>
-                                <td><b>Link to endpoint</b></td>
-                                <td><a href="{{ url }}">{{ url }}</a></td>
-                            </tr>
-                        {% endif %}
-                        <tr>
-                            <td><b>Last accessed</b></td>
-                            <td>{{ "{:%Y-%m-%d %H:%M:%S }".format(rule.last_accessed) }}</td>
-                        </tr>
-                    </table>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <div class="row">
-        <div class="panel panel-default">
-            <div class="panel-heading">
-                Graphs
-            </div>
-            <div class="panel-body">
-                <ul class="nav nav-tabs">
-                    <li><a href="#tph" data-toggle="tab">Time per hour</a></li>
-                    <li><a href="#hph" data-toggle="tab">Hits per hour</a></li>
-                    <li><a href="#tpvpu" data-toggle="tab">Time per version per user</a></li>
-                    <li><a href="#tpvpi" data-toggle="tab">Time per version per ip</a></li>
-                    <li><a href="#tpv" data-toggle="tab">Time per version</a></li>
-                    <li><a href="#tpu" data-toggle="tab">Time per user</a></li>
-                    <li><a href="#nph" data-toggle="tab">Heatmap</a></li>
-                </ul>
-                <div class="tab-content">
-                    <div class="tab-pane fade" id="tph">
-                        <br/>
-                        <div class="col-md-12 col-sm-12">
-                            {{ times_data|safe }}
-                        </div>
-                    </div>
-                    <div class="tab-pane fade" id="hph">
-                        <br/>
-                        <div class="col-md-12 col-sm-12">
-                           {{ hits_data|safe }}
-                        </div>
-                    </div>
-                    <div class="tab-pane fade" id="tpvpu">
-                        <br/>
-                        <div class="col-md-12 col-sm-12">
-                            <embed type="image/svg+xml" src={{ dot_chart_user|safe }}></embed>
-                        </div>
-                    </div>
-                    <div class="tab-pane fade" id="tpvpi">
-                        <br/>
-                        <div class="col-md-12 col-sm-12">
-                            <embed type="image/svg+xml" src={{ dot_chart_ip|safe }}></embed>
-                        </div>
-                    </div>
-                    <div class="tab-pane fade" id="tpv">
-                        <br/>
-                        <div class="col-md-12 col-sm-12">
-                            {{ div_versions|safe }}
-                        </div>
-                    </div>
-                    <div class="tab-pane fade" id="tpu">
-                        <br/>
-                        <div class="col-md-12 col-sm-12">
-                            {{ div_users|safe }}
-                        </div>
-                    </div>
-                    <div class="tab-pane fade" id="nph">
-                        <br/>
-                        <div class="col-md-12 col-sm-12">
-                            {{ div_heatmap|safe }}
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-
-{% endblock %}
