+++ dashboard/templates/endpoint/endpoint_base.html
@@ -1,0 +1,79 @@
+{% extends "main.html" %}
+{% block content %}
+
+    <div class="row">
+        <!-- Advanced Tables -->
+        <div class="panel panel-default">
+            <div class="panel-heading">
+                Summary of endpoint
+            </div>
+            <br/>
+            <div class="panel-body">
+                <div class="table-responsive">
+                    <table class="table table-striped table-bordered table-hover">
+                        <tr>
+                            <td><b>Endpoint</b></td>
+                            <td>{{ rule.endpoint }}</td>
+                        </tr>
+                        <tr>
+                            <td><b>Added since app version</b></td>
+                            <td>{{ rule.version_added }}</td>
+                        </tr>
+                        <tr>
+                            <td><b>Date added to app</b></td>
+                            <td>{{ "{:%Y-%m-%d %H:%M:%S }".format(rule.time_added) }}</td>
+                        </tr>
+                        {% if url %}
+                            <tr>
+                                <td><b>Link to endpoint</b></td>
+                                <td><a href="{{ url }}">{{ url }}</a></td>
+                            </tr>
+                        {% endif %}
+                        <tr>
+                            <td><b>Last accessed</b></td>
+                            <td>{{ "{:%Y-%m-%d %H:%M:%S }".format(rule.last_accessed) }}</td>
+                        </tr>
+                    </table>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <div class="row">
+        <div class="panel panel-default">
+            <div class="panel-heading">
+                Graphs
+            </div>
+            <div class="panel-body">
+                <ul class="nav nav-tabs">
+                    {% if index != 0 %}
+                    <li><a href="{{ url_for('dashboard.result', end=end, index=0) }}">Heatmap</a></li>
+                    {% endif %}
+                    {% if index != 1 %}
+                    <li><a href="{{ url_for('dashboard.result', end=end, index=1) }}">Time per hour</a></li>
+                    {% endif %}
+                    {% if index != 2 %}
+                    <li><a href="{{ url_for('dashboard.result', end=end, index=2) }}">Hits per hour</a></li>
+                    {% endif %}
+                    {% if index != 3 %}
+                    <li><a href="{{ url_for('dashboard.result', end=end, index=3) }}">Time per version per user</a></li>
+                    {% endif %}
+                    {% if index != 4 %}
+                    <li><a href="{{ url_for('dashboard.result', end=end, index=4) }}">Time per version per ip</a></li>
+                    {% endif %}
+                    {% if index != 5 %}
+                    <li><a href="{{ url_for('dashboard.result', end=end, index=5) }}">Time per version</a></li>
+                    {% endif %}
+                    {% if index != 6 %}
+                    <li><a href="{{ url_for('dashboard.result', end=end, index=6) }}">Time per user</a></li>
+                    {% endif %}
+                </ul>
+
+                {% block graph_content %}
+                {% endblock %}
+
+            </div>
+        </div>
+    </div>
+
+{% endblock %}
