+++ dashboard/templates/measurements/measurements_base.html
@@ -36,7 +36,7 @@
                                 {% endfor %}
                                 <td style="text-align: right;">{{ ((record.average*10)|round)/10 }}</td>
                                 <td>
-                                    <a href="{{ url_for('dashboard.result', end=record.endpoint) }}"
+                                    <a href="{{ url_for('dashboard.result', end=record.endpoint, index=0) }}"
                                        class="btn btn-default btn-sm right-align-custom">
                                         Details
                                     </a>
@@ -58,41 +58,23 @@
             </div>
             <div class="panel-body">
                 <ul class="nav nav-tabs">
-                    <li><a href="#tpv" data-toggle="tab">Heatmap</a></li>
-                    <li><a href="#rpepd" data-toggle="tab">Requests per endpoint</a></li>
-                    <li><a href="#etpv" data-toggle="tab">Time per version</a></li>
-                    <li><a href="#etpe" data-toggle="tab">Time per endpoint</a></li>
+                    {% if index != 0 %}
+                        <li><a href="{{ url_for('dashboard.measurements', index=0) }}">Heatmap of number of requests</a></li>
+                    {% endif %}
+                    {% if index != 1 %}
+                        <li><a href="{{ url_for('dashboard.measurements', index=1) }}">Requests per endpoint</a></li>
+                    {% endif %}
+                    {% if index != 2 %}
+                        <li><a href="{{ url_for('dashboard.measurements', index=2) }}">Time per version</a></li>
+                    {% endif %}
+                    {% if index != 3 %}
+                        <li><a href="{{ url_for('dashboard.measurements', index=3) }}">Time per endpoint</a></li>
+                    {% endif %}
                 </ul>
 
-                <div class="tab-content">
-                    <div class="tab-pane fade" id="tpv">
-                        <br/>
-                        <div class="col-md-12 col-sm-12">
-                            {{ heatmap|safe }}
-                        </div>
-                    </div>
-
-                    <div class="tab-pane fade" id="rpepd">
-                        <br/>
-                        <div class="col-md-12 col-sm-12">
-                            <embed type="image/svg+xml" src={{ rpepd|safe }}></embed>
-                        </div>
-                    </div>
-
-                    <div class="tab-pane fade" id="etpv">
-                        <br/>
-                        <div class="col-md-12 col-sm-12">
-                            {{ etpv|safe }}
-                        </div>
-                    </div>
-
-                    <div class="tab-pane fade" id="etpe">
-                        <br/>
-                        <div class="col-md-12 col-sm-12">
-                            {{ etpe|safe }}
-                        </div>
-                    </div>
-                </div>
+                {% block graph_content %}
+                
+                {% endblock %}
             </div>
         </div>
     </div>
