+++ dashboard/templates/measurements/measurements_plotly.html
@@ -1,0 +1,6 @@
+{% extends "measurements/measurements_base.html" %}
+{% block graph_content %}
+    <div class="col-md-12 col-sm-12">
+        {{ graph|safe }}
+    </div>
+{% endblock %}
